cloudrun/gemma3-hf-token/requirements.txt
# requirements.txt
fastapi
uvicorn
python-dotenv

huggingface_hub